% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{Background data selection}
\usage{
background(sp, sp_dat, env)
}
\arguments{
\item{sp}{Species name}

\item{sp_dat}{SpatialPointsDataFrame of species presence records.}

\item{env}{Stack of environmental layers.}
}
\description{
Obtain background information for a species
}
\details{
The function creates an envelope around presence records. All the data within the envelope is considered background.
It will save the envelope as a .tif file in working directory.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
