% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHELSA_data.R
\name{CHELSA_data}
\alias{CHELSA_data}
\title{Download CHELSA bioclimatic data}
\usage{
CHELSA_data(period='current',year=NA,bioclims='all', gcm=NA, ssp=NA)
}
\arguments{
\item{period}{Period in which you want to retrieve data. Possible entries are:
'past', 'current' or 'future'. It is only possible to select one period per run.
Default value is 'current'.}

\item{year}{If you select 'future' as the period, you need to specify which year range.
Possible entries are: '2011-2040', '2041-2070 and/or '2071-2100'.
You can  use a vector to provide more than one entry.}

\item{bioclims}{With this argument it is possible to download only desired bioclimatic
layers, instead of the whole 19. Possible entries are: 'all' (default - retrieve all 19
bioclimatic variables), 'bio1', 'bio2', 'bio3', 'bio4', 'bio5', 'bio6', 'bio7', 'bio8',
'bio9', 'bio10', 'bio11', 'bio12', 'bio13', 'bio14', 'bio15', 'bio16', 'bio17', 'bio18'
and/or 'bio19'. You can use a vector to provide more than one entry.}

\item{gcm}{GCMs to be considered in either past or future periods. Possible entries for
future: 'GFDL-ESM4','IPSL-CM6A-LR','MPI-ESM1-2-HR','MRI-ESM2-0','UKESM1-0-LL'.
Possible entries for past: 'CCSM4','CNRM-CM5','FGOALS-g2','IPSL-CM5A-LR','MIROC-ESM',
'MPI-ESM-P','MRI-CGCM3'. You can use a vector to provide more than one entry.}

\item{ssp}{SSPs for future data. Possible entries are: 'ssp126', 'ssp370', 'ssp585'.
You can use a vector to provide more than one entry.}
}
\description{
This function allows to download bioclimatic data from CHELSA
(https://chelsa-climate.org) considering multiple GCMs, time periods and SSPs.
}
\details{
This function will create a folder entitled 'CHELSA_data'. All the data
downloaded will be stored in this folder. Note that, despite being possible to
retrieve a lot of data at once, it is not recommended to do so, since the data
is very heavy. If you are running models to past and future scenarios, consider
to use GCMs from the same research group, like IPSL, MPI and MRI (see Arguments
section above).
}
\examples{
\dontrun{# download all bioclimatic variables for current period
CHELSA_data()

# download data from future period
CHELSA_data('future', year = '2011-2040', gcm = 'MPI-ESM1-2-HR',
ssp='ssp370', bioclims = c('bio1', 'bio12'))

# download data from past period
CHELSA_data('past', gcm = 'IPSL-CM5A-LR', bioclims = c('bio1','bio12'))}
}
\references{
https://chelsa-climate.org
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
