% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_binaries.R
\name{compare_binaries}
\alias{compare_binaries}
\title{Compare two binary rasters}
\usage{
compare_binaries(r1,r2)
}
\arguments{
\item{r1}{Binary raster from present/current distribution.}

\item{r2}{Binary raster from future/projected distribution.}
}
\description{
This function allows to compare binary rasters to better see patterns.
}
\details{
This function will compare two rasters, returning absence, expansion, retraction and permanence. Absence (value = 1) is a cell where the species is absent in both r1 and r2. Expansion (value = 2) is a cell where species is absent in r1 and present in r2. Retraction (value = 3) is a cell where species is present in r1, but absent in r2. Finally, permanence (value = 4) is a cell where species is present in both r1 and r2.
}
\examples{
\dontrun{r <- compare_binaries(r1,r2)
plot(r, col=c("light gray", "blue", "red", "green"))}
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
