% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.clean.R
\name{data.clean}
\alias{data.clean}
\title{Presence data cleaning wrapper}
\usage{
data.clean(x, r = NULL)
}
\arguments{
\item{x}{data.frame of species records. The output of GBIF_data.}

\item{r}{Optional. A Raster used to delete records in the same cell.}
}
\description{
Data cleaning procedure using CoordinateCleaner and a raster.
}
\details{
Function to clean presence data avoiding autocorrelation.
}
\examples{
# Select species names:
s <- c("Araucaria angustifolia", "Paubrasilia echinata", "Eugenia uniflora")

# Run function:
data <- GBIF_data(s)

# Clean coordinates:
data2 <- data.clean(data)

}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
