% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleoClim_data.R
\name{PaleoClim_data}
\alias{PaleoClim_data}
\title{Download PaleoClim bioclimatic data}
\usage{
PaleoClim_data(tp='current', res=10)
}
\arguments{
\item{tp}{Period in which you want to retrieve data. Possible entries are:
'm2', 'mpwp', 'MIS19', 'LIG', 'LGM', 'HS1', 'BA', 'YDS', 'EH', 'MH', 'LH',
'LH' and/or 'current'. You can select more than one period per run through a vector.
Default value is 'current'.}

\item{res}{Resolution you want to obtain the data. Possible entries are:
'10', '5', '2.5' and/or '30'. You can  use a vector to provide more than one entry.}
}
\description{
This function allows to download bioclimatic data from PaleoClim
(http://www.paleoclim.org) considering multiple time periods.
}
\details{
This function will create a folder entitled 'PaleoClim_data'. All the data downloaded will be stored in this folder. Note that, despite being possible to retrieve a lot of data at once, it is not recommended to do so, since the data is very heavy. Note that PaleoClim is based on CHELSA layers.
}
\examples{
\dontrun{# download bioclimatic variables for YDS period and 10 arc-minutes
PaleoClim_data(c('current', 'YDS'), 10)}
}
\references{
http://www.paleoclim.org
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
