% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedList_status.R
\name{RedList_status}
\alias{RedList_status}
\title{Retrieving IUCN’s RedList Status for Single or Multiple Species}
\usage{
RedList_status(s,k)
}
\arguments{
\item{s}{Species name in a binomial. e.g.: s <- 'Araucaria angustifolia'. s
can also be a vector with multiple species.}

\item{k}{Tolken provided by IUCN.}
}
\description{
This function allows you to retrieve  IUCN’s RedList Status for
a single or multiple species.
}
\details{
Note that to have access to IUCN’s database, you must first get a tolken (an access code). This can be done through a formular in this website: https://apiv3.iucnredlist.org/api/v3/token. It will take a few days to receive this tolken, so don’t be hurry when doing that.
}
\examples{
s <- c('Araucaria angustifolia', 'Ilex paraguariensis')
result <- RedList_status(s,k)
}
\references{
https://apiv3.iucnredlist.org
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
