% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_area.R
\name{w_area}
\alias{w_area}
\title{Weighted Area}
\usage{
w_area(x)
}
\arguments{
\item{x}{Raster of species projection.}
}
\description{
Calculate the area of a species weighting it by species occurrence or environmental suitability.
}
\details{
Function to retrieve area values according to Esser et al. 2019.
}
\examples{
\dontrun{
p1 <- raster(...)
w_area(p1)}
}
\references{
Esser, L. F., Neves, D. M., & Jarenkow, J. A. (2019). Habitat‐specific impacts of climate change in the Mata Atlântica biodiversity hotspot. Diversity and Distributions, 25(12), 1846-1856.
}
\author{
Luíz Fernando Esser (luizesser@gmail.com)
https://luizfesser.wordpress.com
}
